requirements.txt
requests
pandas
python-dotenv
websockets
web3
